Binary search
The binary search algorithm is fast but can be done only if the array is sorted. It
checks the middle element if that is the element that is being searched for. If the search
element is bigger than the middle element, the lower bound is set to the middle element
plus one. If the search element is less than the middle element, the higher bound is set to
the middle element minus one.