Stacks
A stack is a data structure in which only the last inserted element can be removed
and accessed. Think about stacking plates on a table. To get to the
bottom one, you must remove all the other ones on the top. This is a principle known
as last in, first out (`LIFO`). A stack is great because it is fast.