The Deletion in the BST
This algorithm works by first traversing down the tree looking specifically for the node
with the specified value. When the node is found, there are three possible cases:

• Case 1: The node has no children.
This is the simplest case. If the node has no child, return null.
That node has been deleted now.
• Case 2: The node has one child.
If the node has one child, simply return the existing child. That
child has now bubbled up and replaced the parent.
• Case 3: The node has two children.
If the node has two children, either find the maximum of the left
subtree or find the minimum of the right subtree to replace that
node.